# Faster decoder implemented using bufio.Scanner