--- a/decoder.go
+++ b/decoder.go
@@ -18,22 +18,6 @@ type (
 	}
 )
 
-// NewDecoder builds an SSE decoder with a growing buffer.
-// Lines are limited to bufio.MaxScanTokenSize - 1.
-func NewDecoder(in io.Reader) Decoder {
-	return NewDecoderSize(in, 0)
-}
-
-// NewDecoderSize builds an SSE decoder with the specified buffer size (not growing).
-func NewDecoderSize(in io.Reader, bufferSize int) Decoder {
-	d := &decoder{bufio.NewScanner(in), new(bytes.Buffer)}
-	if bufferSize > 0 {
-		d.scanner.Buffer(make([]byte, bufferSize), bufferSize)
-	}
-	d.scanner.Split(scanLinesCR) // See scanlines.go
-	return d
-}
-
 // Decode reads the input stream and interprets the events in it. Any error while reading is  returned.
 func (d *decoder) Decode() (Event, error) {
 	// Stores event data, which is filled after one or many lines from the reader