Less code as I reuse [bufio.Scanner](https://golang.org/pkg/bufio/#Scanner) (with a specific separator function to recognize CRLF / CR / LF). Faster and less memory allocations.

Before:
```
BenchmarkDecodeEmptyEvent-4			 5000000       369 ns/op     72 B/op       3 allocs/op
BenchmarkDecodeEmptyEventWithIgnoredLine-4	 3000000       568 ns/op     90 B/op       5 allocs/op
BenchmarkDecodeShortEvent-4			 3000000       441 ns/op    112 B/op       4 allocs/op
BenchmarkDecode1kEvent-4			 5000000      2748 ns/op   2113 B/op       4 allocs/op
BenchmarkDecode4kEvent-4			  200000      9575 ns/op   8257 B/op       4 allocs/op
BenchmarkDecode8kEvent-4			  100000     19882 ns/op  16449 B/op       4 allocs/op
BenchmarkDecode16kEvent-4			 3000000     40632 ns/op  41059 B/op       6 allocs/op
```
After:
```
BenchmarkDecodeEmptyEvent-4			 5000000       263 ns/op     72 B/op       2 allocs/op
BenchmarkDecodeEmptyEventWithIgnoredLine-4	 5000000       363 ns/op     88 B/op       3 allocs/op
BenchmarkDecodeShortEvent-4			 5000000       304 ns/op    112 B/op       3 allocs/op
BenchmarkDecode1kEvent-4			 1000000      1725 ns/op   2112 B/op       3 allocs/op
BenchmarkDecode4kEvent-4			  200000      5720 ns/op   8256 B/op       3 allocs/op
BenchmarkDecode8kEvent-4			  200000     11410 ns/op  16448 B/op       3 allocs/op
BenchmarkDecode16kEvent-4			 1000000     21835 ns/op  32832 B/op       3 allocs/op
```