@dolmen 2016-10-22
It seems that the `bufio.Scanner.Buffer` method appeared in recent versions of Go. That's why Travis-CI builds are failing. In #8 I'm improvng the Travis-CI config to investigate more.