@dolmen 2016-10-22
I have inlined the constant that were not improving readability. Rebased and push -f.