@alevinval 2016-10-22
I will have to go to sleep and I'll be quite off for some days. I will merge your code once you are happy with it, really nice, I didn't bother to learn how to use the scanner.

You are teaching me a lot.