@dolmen 2016-10-22
`NewDecoderSize()` is now only exported on go >= 1.6.

This is ready for merge.