@alevinval 2016-10-23
LGTM. Merging.